job: fix a warning on newer linux (#765)
This should not actually matter.

+++ src/runtime/job.cpp
@@ -377,7 +377,7 @@ std::string ResourceBudget::format(uint64_t x) {
   int suffix = 0;
   int up = 0;
   static const char *SI[] = {"B", "kiB", "MiB", "GiB", "TiB", "PiB", "EiB" };
-  while (x >= 10000-up) {
+  while (x >= static_cast<uint64_t>(10000-up)) {
     ++suffix;
     up = (x % 1024) >= 512;
     x /= 1024;
